DRTII-715 Remove redundant manifests lookup
